requirements.txt
PyGithub
semgrep==1.144.0
